Correct Hud image metric source shape

+++ src/engine/render/PaintCanvas.cpp
@@ -1233,7 +1233,7 @@ void paintcanvas_ext_dss2_mtx_assign(void *dst, void *src);
 
 void paintcanvas_ext_dss2_ssdraw(void *eng, void *worldM, void *viewM, void *ss);
 
-unsigned short PaintCanvas::GetImage2DWidth(unsigned int index) {
+int PaintCanvas::GetImage2DWidth(unsigned int index) {
     if (index < this->images.count) {
         PCImage2DView *img = (PCImage2DView *) (this->images.data_)[index];
         return img->width;
@@ -1400,7 +1400,7 @@ void PaintCanvas::DrawRectangle(int x, int y, int w, int h) {
     }
 }
 
-unsigned short PaintCanvas::GetImage2DHeight(unsigned int index) {
+int PaintCanvas::GetImage2DHeight(unsigned int index) {
     if (index < this->images.count) {
         PCImage2DView *img = (PCImage2DView *) (this->images.data_)[index];
         return img->height;
